Refactor the common part of gRPC clients
