+++ packages/api/internal/analytics_collector/client.go
@@ -1,76 +1,22 @@
 package analyticscollector
 
 import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"fmt"
-	"os"
-	"regexp"
-	"strings"
-
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/credentials/insecure"
+	"os"
 
 	e2bgrpc "github.com/e2b-dev/infra/packages/shared/pkg/grpc"
 )
 
-var (
-	regex = regexp.MustCompile(`http[s]?://`)
-	host  = regex.ReplaceAllString(os.Getenv("ANALYTICS_COLLECTOR_HOST"), "")
-)
-
-type ClientConnInterface interface {
-	Invoke(ctx context.Context, method string, args any, reply any, opts ...grpc.CallOption) error
-	NewStream(ctx context.Context, desc *grpc.StreamDesc, method string, opts ...grpc.CallOption) (grpc.ClientStream, error)
-	Close() error
-}
+var host = os.Getenv("ANALYTICS_COLLECTOR_HOST")
 
 type Analytics struct {
 	Client     AnalyticsCollectorClient
-	connection ClientConnInterface
-}
-
-func getConnection() (ClientConnInterface, error) {
-	if host == "" {
-		fmt.Println("Analytics collector not set, using dummy connection")
-
-		return &e2bgrpc.DummyConn{}, nil
-	}
-
-	if strings.HasPrefix(host, "localhost") {
-		conn, err := grpc.Dial(host, grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithPerRPCCredentials(&gRPCApiKey{}))
-		if err != nil {
-			return nil, fmt.Errorf("failed to dial: %w", err)
-		}
-
-		return conn, nil
-	}
-
-	systemRoots, err := x509.SystemCertPool()
-	if err != nil {
-		errMsg := fmt.Errorf("failed to read system root certificate pool: %w", err)
-
-		return nil, errMsg
-	}
-
-	cred := credentials.NewTLS(&tls.Config{
-		RootCAs:    systemRoots,
-		MinVersion: tls.VersionTLS13,
-	})
-
-	conn, err := grpc.Dial(host+":443", grpc.WithAuthority(host), grpc.WithTransportCredentials(cred), grpc.WithPerRPCCredentials(&gRPCApiKey{}))
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to dial: %w", err)
-	}
-
-	return conn, nil
+	connection e2bgrpc.ClientConnInterface
 }
 
 func NewAnalytics() (*Analytics, error) {
-	conn, err := getConnection()
+	conn, err := e2bgrpc.GetConnection(host, grpc.WithPerRPCCredentials(&gRPCApiKey{}))
 	if err != nil {
 		return nil, fmt.Errorf("failed to establish GRPC connection: %w", err)
 	}
