+++ packages/api/internal/orchestrator/client.go
@@ -1,75 +1,27 @@
 package orchestrator
 
 import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"fmt"
 	"os"
-	"regexp"
-	"strings"
 
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/credentials/insecure"
 
+	e2bgrpc "github.com/e2b-dev/infra/packages/shared/pkg/grpc"
 	"github.com/e2b-dev/infra/packages/shared/pkg/grpc/orchestrator"
 )
 
 var (
-	regex = regexp.MustCompile(`http[s]?://`)
-	host  = regex.ReplaceAllString(os.Getenv("ORCHESTRATOR_ADDRESS"), "")
+	host = os.Getenv("ORCHESTRATOR_ADDRESS")
 )
-
-type ClientConnInterface interface {
-	Invoke(ctx context.Context, method string, args any, reply any, opts ...grpc.CallOption) error
-	NewStream(ctx context.Context, desc *grpc.StreamDesc, method string, opts ...grpc.CallOption) (grpc.ClientStream, error)
-	Close() error
-}
 
 type GRPCClient struct {
 	Client     orchestrator.SandboxesServiceClient
-	connection ClientConnInterface
-}
-
-func getConnection() (ClientConnInterface, error) {
-	if strings.HasPrefix(host, "localhost") {
-		conn, err := grpc.Dial(
-			host,
-			grpc.WithTransportCredentials(insecure.NewCredentials()),
-			grpc.WithStatsHandler(otelgrpc.NewClientHandler()),
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to dial: %w", err)
-		}
-
-		return conn, nil
-	}
-
-	systemRoots, err := x509.SystemCertPool()
-	if err != nil {
-		errMsg := fmt.Errorf("failed to read system root certificate pool: %w", err)
-
-		return nil, errMsg
-	}
-
-	cred := credentials.NewTLS(&tls.Config{
-		RootCAs:    systemRoots,
-		MinVersion: tls.VersionTLS13,
-	})
-
-	conn, err := grpc.Dial(host, grpc.WithAuthority(host), grpc.WithTransportCredentials(cred))
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to dial: %w", err)
-	}
-
-	return conn, nil
+	connection e2bgrpc.ClientConnInterface
 }
 
 func NewClient() (*GRPCClient, error) {
-	conn, err := getConnection()
+	conn, err := e2bgrpc.GetConnection(host, grpc.WithStatsHandler(otelgrpc.NewClientHandler()))
 	if err != nil {
 		return nil, fmt.Errorf("failed to establish GRPC connection: %w", err)
 	}
